chore: Split Inbox Settings Page Component
Clean up the Inbox settings page component as the logic was growing complex.

- Moved sub pages like collaborators and configuration to new components
- Moved Assignment settings to collaborators tab
